refactor(fe): move grade derivation into AssignmentAccordionTrigger
Grade query and all derived values (submittedCount, problemCount,
scoreText, isDetailActivated) only make sense in the context of the
trigger, so move them there. Dialog state also moves in since it's
only used inside the trigger. AssignmentAccordionItem now just manages
accordion open state and the structural layout.
