+++ apps/frontend/app/(client)/(main)/course/[courseId]/_components/AssignmentAccordionItem.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { assignmentQueries } from '@/app/(client)/_libs/queries/assignment'
 import {
   Accordion,
   AccordionContent,
@@ -8,7 +7,6 @@ import {
 } from '@/components/shadcn/accordion'
 import type { Assignment } from '@/types/type'
 import { ErrorBoundary, Suspense } from '@suspensive/react'
-import { useSuspenseQuery } from '@tanstack/react-query'
 import dayjs from 'dayjs'
 import { useState } from 'react'
 import { toast } from 'sonner'
@@ -29,29 +27,6 @@ export function AssignmentAccordionItem({
   isExercise = false
 }: AssignmentAccordionItemProps) {
   const [isAccordionOpen, setIsAccordionOpen] = useState(false)
-  const [isAssignmentDialogOpen, setIsAssignmentDialogOpen] = useState(false)
-
-  const { data: grades } = useSuspenseQuery(
-    assignmentQueries.grades({
-      courseId: Number(assignment.group.id),
-      isExercise
-    })
-  )
-  const grade = grades.find((g) => g.id === assignment.id)
-
-  const submittedCount = grade?.submittedCount ?? 0
-  const problemCount = grade?.problemCount ?? assignment.problemCount
-
-  let scoreText = '- / -'
-  let isDetailActivated = false
-
-  if (!isExercise && grade) {
-    const userScore = grade.userAssignmentFinalScore ?? '-'
-    scoreText = `${userScore} / ${grade.assignmentPerfectScore}`
-    isDetailActivated =
-      grade.userAssignmentFinalScore !== null &&
-      dayjs().isAfter(dayjs(assignment.endTime))
-  }
 
   return (
     <AssignmentAccordionProvider
@@ -70,15 +45,7 @@ export function AssignmentAccordionItem({
           value={assignment.id.toString()}
           className="group border-b-0"
         >
-          <AssignmentAccordionTrigger
-            hasStarted={grade !== undefined}
-            submittedCount={submittedCount}
-            problemCount={problemCount}
-            scoreText={scoreText}
-            isDetailActivated={isDetailActivated}
-            isAssignmentDialogOpen={isAssignmentDialogOpen}
-            onAssignmentDialogChange={setIsAssignmentDialogOpen}
-          />
+          <AssignmentAccordionTrigger />
           <AccordionContent className="-mb-4 w-full">
             {isAccordionOpen && (
               <ErrorBoundary
@@ -92,7 +59,9 @@ export function AssignmentAccordionItem({
               >
                 <Suspense
                   fallback={
-                    <AssignmentProblemListSkeleton count={problemCount} />
+                    <AssignmentProblemListSkeleton
+                      count={assignment.problemCount}
+                    />
                   }
                 >
                   <AssignmentProblemList />
